requirements.txt
beautifulsoup4==4.10.0
chromadb==0.5.0
gpt-index==0.5.8
langchain==0.2.1
langchain-chroma==0.1.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langgraph==0.0.61
langsmith==0.1.67
llama-index==0.5.8
lockfile==0.12.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.5
Pillow==9.0.1
streamlit==1.35.0
streamlit-chat==0.0.2.2
